feat: default to float16 leaf storage
Switch init()'s leaf_dtype default from float32 to float16.

Adapt the test suite without baking in float16:
- assert_close_matrices upcasts inputs to float64 before the norm, so
  reduced-precision dtypes (rejected by scipy's norm) and booleans work.
- add condf(array, on_32, on_16) helper that selects a tolerance by the
  leaf storage precision, so equivalence/precision tests loosen to the
  rounding floor under float16 and revert to their tight tolerances
  automatically when leaf storage goes back to float32.
- TestLeafDtype.test_dtypes_and_scale reads the default off init()'s
  signature instead of hardcoding float32.

+++ src/bartz/mcmcstep/_state.py
@@ -537,7 +537,7 @@ def init(
     num_trees: int,
     p_nonterminal: Float32[ArrayLike, ' d_minus_1'],
     leaf_prior_cov_inv: FloatLike | Float[ArrayLike, 'k k'],
-    leaf_dtype: DTypeLike = jnp.float32,
+    leaf_dtype: DTypeLike = jnp.float16,
     error_cov_df: FloatLike | None = None,
     error_cov_scale: FloatLike | Float[ArrayLike, 'k k'] | None = None,
     error_scale: Float32[ArrayLike, ' n'] | Float32[ArrayLike, 'k n'] | None = None,
